refactor(ci-docker): add coverage workflows and split web utilities into focused packages
Align CI/CD with current coverage baseline, modernize Docker/Compose runtime behavior, and reduce directory crowding by extracting env, TLS, request, and session cookie utilities without changing app behavior.

Made-with: Cursor
