+++ .gitignore
@@ -5,3 +5,4 @@ ui/static/images/*.jpg
 ui/static/images/*.gif
 tls/*.pem
 .env
+coverage.out
