+++ cmd/web/envfile_compat.go
@@ -1,0 +1,7 @@
+package main
+
+import "github.com/aspandyar/forum/internal/config/envfile"
+
+func LoadEnvFromFile(filename string) error {
+	return envfile.Load(filename)
+}
